+++ Changelog.md
@@ -1,6 +1,8 @@
 ### 0.8.15 (unreleased)
 
 Language Features:
+* General: Add `E.selector` for a non-anonymous event `E` to access the 32-byte selector topic.
+* General: Errors and Events allow qualified access from other contracts.
 
 
 Compiler Features: